create SubareaButton class to implement toggle show/hide subarea panel. CreateSubareaPanels should hand over aforementioned responsibility to SubareaButton class. Create getDimensions method in MapPanel to determine the number of r,c needed for the game map grid
