+++ src/main/java/client/StrainXMain.java
@@ -1,5 +1,6 @@
 package client;
 
+import gamemusic.AudioPlayer;
 import ui.gui.GUIEntry;
 import java.io.IOException;
 
@@ -9,6 +10,5 @@ public class StrainXMain {
 
 //        UIEnterGame.displayEnterGame();
         GUIEntry.start();
-
     }
 }
