+++ src/main/java/ui/gui/GUIEntry.java
@@ -14,7 +14,6 @@ import java.io.IOException;
 
 public class GUIEntry {
     public static void start() throws IOException {
-
         // TEST StatsPanel
         JFrame jFrame = new JFrame();
 
@@ -27,6 +26,6 @@ public class GUIEntry {
         jFrame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
         jFrame.pack();
         jFrame.setVisible(true);
-        jFrame.add(wrapperPanel);
+//        jFrame.add(wrapperPanel);
     }
 }
